MVP_Chris/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-dotenv
pandas
scikit-learn
numpy
psycopg2-binary
pytest
httpx
greenlet
reportlab
